classga otkazildi

+++ asosiy/views.py
@@ -1,33 +1,49 @@
+from django.views import View
 from django.shortcuts import render, redirect
 from django.contrib.auth import authenticate, login, logout
 from .models import *
 from .forms import *
 
-def home(request):
-    if request.user.is_authenticated:
-        if request.method == "POST":
-            form =KundalikForm(request.POST)
-            if form.is_valid():
-                plan = form.save(commit=False)
-                plan.foydalanuvchi = request.user
-                plan.save()
+class HomeView(View):
+    def post(self, request):
+        form = KundalikForm(request.POST)
+        if form.is_valid():
+            plan = form.save(commit=False)
+            plan.foydalanuvchi = request.user
+            plan.save()
+        return redirect("/todo/")
+    def get(self, request):
+        if request.user.is_authenticated:
+            data = {
+                "kundaliklar":Kundalik.objects.filter(foydalanuvchi=request.user),
+                "kundalik":KundalikForm()
+            }
+            return render(request, 'todo.html', data)
+        return redirect("/")
+
+class Todo_ochirView(View):
+    def get(self, request, son):
+        if request.user.is_authenticated:
+            todo = Kundalik.objects.get(id=son)
+            if todo.foydalanuvchi == request.user:
+                todo.delete()
             return redirect("/todo/")
-        data = {
-            "kundaliklar":Kundalik.objects.filter(foydalanuvchi=request.user),
-            "kundalik":KundalikForm()
-        }
-        return render(request, 'todo.html', data)
-    return redirect("/")
 
-def todo_ochirish(request, son):
-    todo = Kundalik.objects.get(id=son)
-    todo.delete()
+        return redirect("/")
 
-    return redirect("/")
 
 # 1
-def todo_edit(request, son):
-    if request.method == "POST":
+class Todo_editView(View):
+    def get(self, request, son):
+        if request.user.is_authenticated:
+            data = {
+                "todo": Kundalik.objects.get(id=son),
+                "st": ["Boshlanadi", "Boshlandi", "Bajarildi"]
+            }
+            return render(request, "todo_edit.html", data)
+        return redirect("/todo/")
+
+    def post(self, request, son):
         Kundalik.objects.filter(id=son).update(
             sarlavha = request.POST.get('s'),
             muddat = request.POST.get('m'),
@@ -35,23 +51,31 @@ def todo_edit(request, son):
             status = request.POST.get('status')
         )
         return redirect("/")
-    data = {
-        "todo":Kundalik.objects.get(id=son),
-        "st":["Boshlanadi", "Boshlandi", "Bajarildi"]
-    }
-    return render(request, "todo_edit.html", data)
 
-def loginview(request):
-    if request.method == "POST":
+
+class LoginView(View):
+    def post(self, request):
         user = authenticate(username=request.POST.get('l'),
-                     password=request.POST.get('p'))
+                            password=request.POST.get('p'))
         if user is None:
             return redirect("/")
         login(request, user)
         return redirect("/todo/")
-    return render(request, 'login.html')
+    def get(self, request):
+        return render(request, 'login.html')
 
 
-def logoutview(request):
-    logout(request)
-    return redirect("/")
+
+class LogoutView(View):
+    def get(self, request):
+        logout(request)
+        return redirect("/")
+
+def register(request):
+    if request.method == "POST" and request.POST.get('p') == request.POST.get('cp'):
+        User.objects.create_user(
+            username = request.POST.get('l'),
+            password = request.POST.get('p')
+        )
+        return redirect("/")
+    return render(request, 'register.html')
